requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
PyMuPDF==1.24.10
sentence-transformers==3.1.1
chromadb==0.5.15
rank-bm25==0.2.2
groq==0.11.0
google-generativeai==0.8.3
pydantic==2.9.2
httpx==0.27.2
python-dotenv==1.0.1
ragas==0.1.21
datasets==2.20.0
langchain-openai==0.1.25
openai==1.109.1
pinecone-client==3.2.2